backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-limiter==3.5.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
firebase-admin==6.3.0
pyjwt==2.8.0
python-dotenv==1.0.0
eventlet==0.35.2
apscheduler==3.10.4
gunicorn==21.2.0
bleach==6.1.0
pillow==10.2.0
python-magic==0.4.27
celery==5.3.6